Issue #5: Changing package.json after installing codacy plugin
This commit changes the package.json as a result of installing the
dev depencencies for Codacy as described at
https://github.com/codacy/node-codacy-coverage#installation

In addition, the section for the keywords.

Finally, the script "npm test" has been changed to what has been
documented to be used with the following:

https://github.com/codacy/node-codacy-coverage#istanbul

* istanbul Code Coverage
* Mocha Runner
* LCov reporter

Locally, this is resulting in an error for sending the data to codacy.